Reconstruction
It is the configurations I am using now, which are on archlinux and Hyprland desktop env. Due to the discovery of two potential information leakage security risks in the previous warehouse, the stop loss, it was removed. Now, with tweaks and other improvements, the repository has been rebuilt, still under the GPL-3.0 License and public.
1. Added security warnings for application configuration, terminal installation and its preconditions;
2. Added a standby hotkey to Kitty to prevent users who do not have alacritty installed from failing to start the terminal after the application configuration;
3. Added the content of the SDDM configuration file, and at the same time made a more detailed explanation of the instructions for new users;
4. Supplemented with an effect file of related configuration (has processed);
5. The previous information is sensitive or deleted.
This is how this repository changed today (UTC 07/30/2024).